# Conflicts:
#	app/src/main/java/ru/tinkoff/tinkoffer/presentation/screen/home/components/ProposalElement.kt

+++ app/src/main/java/ru/tinkoff/tinkoffer/data/models/projects/request/CreateProjectDto.kt
@@ -6,6 +6,6 @@ import kotlinx.serialization.Serializable
 data class CreateProjectDto(
     val name: String,
     val schedule: String,
-    val votesPerPeriod: Long,
-    val votesRefreshPeriodDays: Long,
+    val votesPerPeriod: Int,
+    val votesRefreshPeriodDays: Int,
 )
